% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_crowns_to_las.R
\name{add_crowns_to_las}
\alias{add_crowns_to_las}
\title{Add Crown IDs to LAS Files}
\usage{
add_crowns_to_las(
  proj_dir,
  acq = NULL,
  num_cores = 1L,
  chunk_buf = NULL,
  crown_dir = NULL,
  crown_pattern = "lmfws2",
  crown_ext = ".gpkg$",
  filter = FALSE
)
}
\arguments{
\item{proj_dir}{Project directory containing the input LAS files and the output directories.}

\item{acq}{Acquisition name. If NULL (default), the basename of the \code{proj_dir} will be used as the acquisition name.}

\item{num_cores}{Number of cores to use for parallel processing. Default is 1.}

\item{chunk_buf}{Chunk buffer size. If NULL (default), it is auto-set to 5\% of tile_size.}

\item{crown_dir}{Directory containing the crown shapefiles. Default is NULL.}

\item{crown_pattern}{Pattern to match the crown files. Default is 'lmfws2'.}

\item{crown_ext}{File extension for the crown files. Default is '.gpkg$'.}

\item{filter}{Logical to decide whether to apply any filtering while adding tree crowns. Default is FALSE.}
}
\value{
Creates new LAS files with tree crown IDs attached as attribute.
}
\description{
This function processes the LAS files in a catalog to add tree crown IDs based on the supplied crown vectors or rasters
}
