% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape_metrics.R
\name{alphashape_metrics}
\alias{alphashape_metrics}
\title{Generate Crown Structural Metrics}
\usage{
alphashape_metrics(las, snapshot = FALSE, image_output = NULL, acq = NULL)
}
\arguments{
\item{las}{single tree las object}

\item{snapshot}{Logical indicating whether to generate snapshots of the 3D alphashape (default: \code{FALSE})}

\item{image_output}{Path to directory where the snapshots will be saved (default: \code{NULL})}

\item{acq}{Identifier to include in the filename of the snapshot (default: \code{NULL})}
}
\value{
returns an sf object with 3D convex hull metrics
}
\description{
Generate Crown Structural Metrics
}
\examples{
\dontrun{
las <- lidR::filter_poi(tree_las, treeID == 17)
ashape <- alphashape_metrics(las)
}
}
