% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_cindex.R
\name{apa_cindex}
\alias{apa_cindex}
\title{Calculate Area Potentially Avaliable Indicies}
\usage{
apa_cindex(ttops, comp_input = "Zmax")
}
\arguments{
\item{ttops}{tree tops object; expects an X.detected, Y.detected column, typically from silvtools::tree_matching() function}

\item{comp_input}{string value indicating competition index input specifically related to size; traditionally DBH/Height, crown volume etc}
}
\value{
returns original ttops dataset (sf/dataframe) with new cindex column
}
\description{
Calculate Area Potentially Avaliable Indicies
}
\examples{
\dontrun{
c <- apa_cindex(ttops, comp_input = 'Zmax')
}
}
