% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_treeid_to_las.R
\name{apply_treeid}
\alias{apply_treeid}
\title{Applies tree IDs to LAS file based on crown intersections}
\usage{
apply_treeid(chunk, crowns, filter = TRUE, buf = 5)
}
\arguments{
\item{chunk}{A LAS file chunk to be processed.}

\item{crowns}{A set of crowns represented as an sf object.}

\item{filter}{Logical indicating whether to filter out points that do not correspond to a tree}

\item{buf}{Buffer distance to use when selecting crowns that intersect with the chunk}
}
\value{
A LAS file with an additional attribute called "treeID" that contains
the ID of the tree to which each point belongs.
}
\description{
This function reads in a LAS file chunk and intersects it with a set of crowns
to assign tree IDs to each point in the LAS file that belongs to a tree. The
resulting LAS file will have an additional attribute called "treeID" that
contains the ID of the tree to which each point belongs.
}
\examples{

lasfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las <- readLAS(lasfile)
norm_las <- normalize_height(las)




}
