% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_chm_crowns.R
\name{approximate_chm_crowns}
\alias{approximate_chm_crowns}
\title{Segment tree crowns from a CHM}
\usage{
approximate_chm_crowns(
  proj_dir,
  chm_dir = NULL,
  vector_output = NULL,
  raster_output = NULL,
  crown_methods = c("fixed", "auto", "variable"),
  hmin = 1,
  crown_height_threshold = 0.25,
  vis = FALSE,
  chm_ext = "smooth",
  window_size = NULL
)
}
\arguments{
\item{proj_dir}{A string specifying the project directory.}

\item{chm_dir}{A string specifying the raster output directory. Default is NULL.}
}
\value{
A list containing the results of each segmentation method.
}
\description{
This function segments tree crowns using a canopy height model (CHM) based on the tree tops located earlier.
}
\examples{
\dontrun{
approximate_chm_crowns("path/to/project")
}
}
