% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_chm_treetops.R,
%   R/approximate_trees.R
\name{approximate_chm_treetops}
\alias{approximate_chm_treetops}
\title{Approximate tree tops from a CHM}
\usage{
approximate_chm_treetops(
  proj_dir,
  chm_dir = NULL,
  vector_output = NULL,
  fixed_window = TRUE,
  auto_window = TRUE,
  variable_window = TRUE,
  fix_ws = 2,
  hmin = 2,
  mod = 0.07
)

approximate_chm_treetops(
  proj_dir,
  chm_dir = NULL,
  vector_output = NULL,
  fixed_window = TRUE,
  auto_window = TRUE,
  variable_window = TRUE,
  fix_ws = 2,
  hmin = 2,
  mod = 0.07
)
}
\arguments{
\item{proj_dir}{A string specifying the project directory.}

\item{chm_dir}{A string specifying the raster output directory. Default is NULL.}

\item{vector_output}{A string specifying the vector output directory. Default is NULL.}

\item{fixed_window}{A logical indicating whether to perform fixed window tree detection. Default is TRUE.}

\item{auto_window}{A logical indicating whether to perform automatic window tree detection. Default is TRUE.}

\item{variable_window}{A logical indicating whether to perform variable window tree detection. Default is TRUE.}

\item{fix_ws}{A numeric specifying the fixed window size for tree detection. Default is 2.}

\item{hmin}{A numeric specifying the minimum tree height. Default is 2m.}

\item{mod}{A numeric modifier for the variable window size function. Default is 0.07.}

\item{chm_ext}{A string related to the desired chm file selection pattern (default smooth) if multiple options; user chooses}
}
\value{
A list containing the results of each tree detection method.

A list containing the results of each tree detection method.
}
\description{
This function approximates tree locations using a canopy height model (CHM).
It performs tree detection with fixed, automatic, and variable window sizes.

This function approximates tree locations using a canopy height model (CHM).
It performs tree detection with fixed, automatic, and variable window sizes.
}
\examples{
\dontrun{
approximate_chm_treetops("path/to/project")
}
approximate_chm_treetops("path/to/project")
}
