% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_convex_vol_py.R
\name{calc_convex_vol_py}
\alias{calc_convex_vol_py}
\title{Run PyInstaller-packaged Python Script to Calculate Crown Volumes (convex)}
\usage{
calc_convex_vol_py(input_parquet, output_parquet, py_executable)
}
\arguments{
\item{input_parquet}{Path to the input Parquet file containing 3D points.}

\item{output_parquet}{Path to the output Parquet file where the results will be saved.}

\item{py_executable}{Path to the PyInstaller-packaged Python executable.}
}
\value{
A character vector containing the output from the Python script execution.
}
\description{
This function runs a PyInstaller-packaged Python script to calculate the convex volume of 3D points from a Parquet file.
}
\examples{
\dontrun{
input_path <- "G:/calc_concave_vol/tree_las.parquet"
output_path <- "G:/calc_concave_vol/tree_las_output_2.parquet"
executable_path <- "G:/calc_concave_vol/dist/calc_concave_volume/calc_concave_volume.exe"
result <- calc_convex_vol_py(input_path, output_path, executable_path)
print(result)
}
}
