% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_twi.R
\name{calc_twi}
\alias{calc_twi}
\title{DEM to Topographic Wetness Index}
\usage{
calc_twi(dem_file, proj_dir)
}
\arguments{
\item{dem_file}{location of elevation model raster}

\item{proj_dir}{directory where /output/raster/topography dir will be created and outputs stored}
}
\value{
saves a median filter, slope, filled dem, flow accumulation raster, and TWI raster in out_dir
}
\description{
DEM to Topographic Wetness Index
}
\examples{
\dontrun{
proj_dir <- 'H:/Quesnel_2022/blocks/CT1'
dem_file <- file.path(proj_dir, 'output/raster/dtm/ULS22_CT1_dtm_tin_0.25m.tif')
}

}
