% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_las_process.R
\name{configure_las_process}
\alias{configure_las_process}
\title{Lidar Processing Switches}
\usage{
configure_las_process()
}
\description{
A Shiny application for setting lidar processing switches.
}
\details{
Configure Lidar Processing Switches

This Shiny application allows users to set processing switches for lidar data processing.
It provides an interactive user interface to select various options and displays the selected options in a table.
}
\examples{
\dontrun{
  # Run the Shiny app
  configure_las_process()
}
}
\seealso{
\code{\link[shiny]{shinyApp}}
}
