% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_multi_to_single_polygons.R
\name{convert_multi_to_single_polygons}
\alias{convert_multi_to_single_polygons}
\title{Convert MultiPolygons to Single Polygons}
\usage{
convert_multi_to_single_polygons(polygons, fill_holes = TRUE)
}
\arguments{
\item{polygons}{a sf object of polygon geometries.}

\item{fill_holes}{a logical indicating whether to fill any holes in the
resulting polygon (default is TRUE).}
}
\value{
A sf object of single polygon geometries.
}
\description{
This function takes in a set of 'sf' polygon objects and separates them into
single polygon objects and multi-polygon objects. The function then takes
the largest polygon from each multi-polygon and fills any holes in the
resulting polygon (optional). The final output is a set of single polygon
objects.
}
\examples{

# Load example data
library(sf)
library(nngeo)

# Load example MULTIPOLYGON tree crown dataset
multi_crowns <- st_read(system.file("extdata", "multi_crowns.gpkg", package = "silvtools"))

# Apply function to convert example data to only largest single polygons
crowns <- convert_multi_to_single_polygons(polygons = multi_crowns, fill_holes = TRUE)

}
