% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_animation.R
\name{create_animation}
\alias{create_animation}
\title{Create Animation
Animates solar irradiance tifs (or any images) into gif and saves locally}
\usage{
create_animation(proj_dir, fps = 10, label = TRUE, filetype = ".tif$")
}
\arguments{
\item{proj_dir}{Directory wher eproject is stored; expects an /output/raster/irradiance from solar_simulator, creates a /gifs output folder for saving}

\item{fps}{Chosen frames per second for final gif; change depending on number of images default is 10 fps}

\item{label}{Logical; TRUE/FALSE. Determines if final animation is annotated with YYYY-MM-DD HH:MM:SS time stamp for each frame}

\item{filetype}{File type of animation images; default assumes raster in .tif format}
}
\description{
Create Animation
Animates solar irradiance tifs (or any images) into gif and saves locally
}
