% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_buffer}
\alias{create_buffer}
\title{Create Buffer Around Points in an sf DataFrame}
\usage{
create_buffer(chunk, df, distance)
}
\arguments{
\item{chunk}{An sf dataframe containing the chunk of points to be buffered.}

\item{df}{An sf dataframe containing the original set of points.}

\item{distance}{Numeric. The buffer distance to be applied around each point in the chunk.}
}
\value{
An sf dataframe with an additional column \code{buffer_point} indicating which points
fall within the buffer and are not in the original chunk.
}
\description{
This function creates a buffer around each chunk of points in a given sf dataframe,
dissolves borders between the buffered points, and updates the dataframe to indicate
which points fall within the buffer but are not in the original chunk.
}
\examples{
\dontrun{
  library(sf)
  # Create example data
  df <- st_as_sf(data.frame(treeID = 1:10, geometry = st_sfc(st_point(c(0, 0)), st_point(c(1, 1)),
  st_point(c(2, 2)), st_point(c(3, 3)), st_point(c(4, 4)), st_point(c(5, 5)), st_point(c(6, 6)),
  st_point(c(7, 7)), st_point(c(8, 8)), st_point(c(9, 9)))))
  chunk <- df[1:5, ]
  # Apply buffer
  result <- create_buffer(chunk, df, 2)
  print(result)
}

}
