% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_mask.R
\name{crown_mask}
\alias{crown_mask}
\title{Watershed segmentation of crowns with threshold}
\usage{
crown_mask(
  chunk,
  ttops = NULL,
  chm_res = 0.25,
  crown_height_threshold = 0.5,
  ws = 2,
  hmin = 2,
  uniqueness = "incremental",
  vis = FALSE
)
}
\arguments{
\item{chunk}{lasCatalog chunk (catalog_apply) or las file loaded with readLAS or CHM loaded as SpatRaster}

\item{ttops}{optional treetops argument, if provided lmf process is skipped}

\item{chm_res}{desired resolution of chm for segmentation}

\item{crown_height_threshold}{desired threshold of treetop height to mask crown}

\item{ws}{local maxima finding window size}

\item{hmin}{minimum height of local maxima to identify as treetops}

\item{uniqueness}{see ?lidR::locate_trees argument defines how tree IDs are determined}

\item{vis}{LOGICAL; choose whether or not to plot result (default = FALSE)}
}
\description{
Height limited watershed segmentation; takes a las/chunk/chm and performs watershed segmentation
if treetops are provided it will use those; function masks crowns by crown_height_threshold defined as
percentage of treetop height; this will determine if pixels are included in final crown (eg if CHM pixel less than 50\% of tree height; exclude)
}
