% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_setup_als.R
\name{dir_setup_als}
\alias{dir_setup_als}
\title{Setup ALS Directories}
\usage{
dir_setup_als(proj_dir)
}
\arguments{
\item{proj_dir}{Main directory for processing or site}
}
\value{
Generate standardized folders for inputs and outputs
associated with ALS data

Multilevel, /input, /output, /doc, /script

Doc intended to store relevant literature, site info, data collection info

Script intended to house R code and other code used to process input dataset

input/raster, input/lidar, input/vector, and associated output folders generated
}
\description{
Setup ALS Directories
}
\examples{
proj_dir <- 'D:/BC/MKRF/Mixedwoods_trial_2022'
dir_setup_als(proj_dir)
}
