% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chm.R
\name{generate_chm}
\alias{generate_chm}
\title{Generate Canopy Height Model (CHM)}
\usage{
generate_chm(
  proj_dir,
  chm_res = 1,
  num_cores = 1L,
  chunk_buf = NULL,
  acq = NULL,
  clamp = FALSE,
  min_clamp = 0,
  max_clamp = 50
)
}
\arguments{
\item{proj_dir}{Project directory containing the normalized LAS files.}

\item{chm_res}{CHM resolution in meters. Default is 1.}

\item{num_cores}{Number of cores for parallel processing. Default is 1.}

\item{chunk_buf}{Chunk buffer size in meters. Default is NULL (auto-set to 5\% of tile_size).}

\item{acq}{Acquisition name. Default is NULL.}

\item{clamp}{Logical switch to clamp CHM values. Default is FALSE.}

\item{min_clamp}{Minimum value to clamp CHM at. Default is 0.}

\item{max_clamp}{Maximum value to clamp CHM at. Default is 50.}
}
\value{
A SpatRaster representing the CHM.
}
\description{
This function generates a Canopy Height Model (CHM) from normalized LAS files.
}
