% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dtm.R
\name{generate_dtm}
\alias{generate_dtm}
\title{Generate Digital Terrain Model (DTM)}
\usage{
generate_dtm(
  proj_dir,
  dtm_res,
  num_cores = 1L,
  chunk_buf = NULL,
  dtm_algorithm = NULL,
  acq = NULL
)
}
\arguments{
\item{proj_dir}{Project directory containing the classified LAS files.}

\item{dtm_res}{DTM resolution in meters.}

\item{num_cores}{Number of cores for parallel processing. Default is 1.}

\item{chunk_buf}{Chunk buffer size in meters. Default is NULL (auto-set to 5\% of tile_size).}

\item{dtm_algorithm}{Custom DTM algorithm. Default is NULL (TIN).}

\item{acq}{Acquisition name. Default is NULL.}
}
\value{
A SpatRaster representing the DTM.
}
\description{
This function generates a Digital Terrain Model (DTM) from classified LAS files.
}
