% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphashape_metrics.R
\name{get_alphashape_metrics}
\alias{get_alphashape_metrics}
\title{Convert LAS file to Alphashape and compute crown metrics}
\usage{
get_alphashape_metrics(chunk, RGB = FALSE)
}
\arguments{
\item{chunk}{an object of class LAS or a catalog chunk which has points attributed with treeID values, typically computed from lidR::segment_trees.}
}
\value{
A data frame with columns treeID, X, Y, Zmax, Zq999, Zq99, Z_mean, n_points, vol_convex, vol_concave, vol_a05, CV_Z, CRR.
}
\description{
The get_alphashape_metrics function takes a tree-attributed LAS object and converts it to a 3D alphashape.
Crown metrics such as X, Y, Zmax, Zq999, Zq99, Z_mean, n_points, vol_convex, vol_concave, vol_a05, CV_Z, and CRR are then computed for each tree.
}
\examples{
\dontrun{
library(alphashape3d)
library(lidR)
las <- readLAS(system.file("extdata", "uls.laz", package = "silvtools"))
tree_las <- segment_trees(las, dalponte2016(chm = rasterize_canopy(las, res = 1, p2r()),
treetops = locate_trees(las, lmf(ws = 5, hmin = 5))))
ashape_df <- get_alphashape_metrics(tree_las, prog_bar = TRUE)
}
}
