% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crown_attributes.R
\name{get_crown_attributes}
\alias{get_crown_attributes}
\title{Calculate crown attributes for a set of tree points}
\usage{
get_crown_attributes(X, Y, Z, pb = NULL)
}
\arguments{
\item{X}{Numeric vector, x-coordinates of tree points}

\item{Y}{Numeric vector, y-coordinates of tree points}

\item{Z}{Numeric vector, z-coordinates of tree points}

\item{pb}{progress bar object, used for progress tracking (optional)}
}
\value{
A data frame with columns for each crown attribute
}
\description{
This function calculates various crown attributes for a set of 3D tree points including:
crown height (Zmax, Zq999, Zq99, Z_mean), crown size (n_points, vol_convex, vol_concave, vol_a05), and crown complexity (CV_Z, CRR)
}
\examples{
\dontrun{
# For example usage, see the parent function get_alphashape_metrics()
}

}
