% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solar_pos.R
\name{get_solar_pos}
\alias{get_solar_pos}
\title{Get solar positions}
\usage{
get_solar_pos(start_date, end_date, interval, lat, lon)
}
\arguments{
\item{start_date}{Start date/time (POSIXct format)}

\item{end_date}{End date/time (POSIXct format)}

\item{interval}{Observation interval one of: '10 min', '30 min', 'hour'}

\item{lat}{Numeric latitude value (WGS84)}

\item{lon}{Numeric longitude value (WGS84)}
}
\value{
Returns a dataframe of solar positions for given location and time period
}
\description{
Get solar positions
}
\examples{
\dontrun{
start_date <- as.POSIXct("2022-05-15 00:00:00", tz = "America/Los_Angeles")
end_date <- as.POSIXct("2022-09-15 00:00:00", tz = "America/Los_Angeles")
interval <- '10 min'
lat <- 53.371759251761766
lon <- -122.76808019195623

solar_pos <- get_solar_pos(start_date, end_date, interval, lat, lon)
}
}
