% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_size.R
\name{get_tile_size}
\alias{get_tile_size}
\title{Get Tile Size from a LAScatalog}
\usage{
get_tile_size(ctg)
}
\arguments{
\item{ctg}{A LAScatalog object.}
}
\value{
The size of the middle tile, rounded to the nearest whole number.
}
\description{
This internal function calculates the size of the middle tile in a LAScatalog.
The size is rounded to the nearest whole number.
}
\keyword{internal}
