% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heygi_cindex.R
\name{heygi_cindex}
\alias{heygi_cindex}
\title{Calculate Heygi Style Competition Indicies}
\usage{
heygi_cindex(ttops, comp_input = "vol_convex", maxR = 6, quiet = FALSE)
}
\arguments{
\item{ttops}{tree tops object}

\item{comp_input}{string value indicating competition index input; traditionally DBH/Height, we test crown structural metrics}

\item{maxR}{radius of sphere of influence; treetops within this radius will be included in the calculation of cindex}

\item{quiet}{logical; if TRUE, suppresses progress messages}
}
\value{
returns original ttops dataset (sf/dataframe) with new cindex column
}
\description{
Calculate Heygi Style Competition Indicies
}
\examples{
\dontrun{

lasfile <- system.file("extdata", "MixedConifer.laz", package = "lidR")
las <- readLAS(lasfile)
norm_las <- normalize_height(las, tin())
chm <- rasterize_canopy(norm_las, p2r(subcircle = 0.25), res = 1)
ttops <- locate_trees(chm, algorithm = lmf(ws = 2, hmin = 5))

c <- heygi_cindex(ttops, comp_input = 'Z', maxR = 6)
}
}
