% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcwatershed.R
\name{mcwatershed}
\alias{mcwatershed}
\title{Marker Controlled Watershed Segmentation}
\usage{
mcwatershed(chm, treetops, th_tree = 2, tree_id = "treeID")
}
\arguments{
\item{chm}{\itemize{
\item Canopy height model SpatRast
}}

\item{treetops}{\itemize{
\item Local Maxima (Markers)
}}

\item{th_tree}{\itemize{
\item Minimum height threshold for segmented trees (Default is 2m)
}}

\item{tree_id}{\itemize{
\item String of field refering to tree ID attribute associated with treetops
}}
}
\value{
Returns SpatRast of segmented trees by treeID
}
\description{
Marker Controlled Watershed Segmentation
}
\examples{
\dontrun{
mcwatershed(chm, treetops, th_tree = 2, tree_id = 'treeID')()
}
}
