% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ground_classification.R
\name{perform_ground_classification}
\alias{perform_ground_classification}
\title{Perform Ground Classification on a LAScatalog}
\usage{
perform_ground_classification(
  proj_dir,
  tile_size = NULL,
  chunk_buf = NULL,
  acq = NULL,
  ground_algorithm = NULL,
  num_cores = 1L,
  index_tiles = TRUE,
  output_laz = TRUE
)
}
\arguments{
\item{proj_dir}{The project directory where the LAS files are stored.}

\item{tile_size}{The size of each tile for processing. Default is NULL, which will use the size of the middle tile in the catalog.}

\item{chunk_buf}{The buffer around each chunk. Default is NULL, set to 5\% of tile_size.}

\item{acq}{The name of the acquisition, default is NULL which uses the basename of the project directory.}

\item{ground_algorithm}{The ground classification algorithm to use. Default is NULL which uses the CSF algorithm.}

\item{num_cores}{The number of cores for parallel processing. Default is 1.}

\item{index_tiles}{Boolean, whether to index the classified tiles. Default is TRUE.}

\item{output_laz}{Logical, should the output be in LAZ format? Default is TRUE.}
}
\value{
A LAScatalog containing the ground classified points.
}
\description{
Classify the ground points in a LAScatalog using a specified algorithm.
Automatically sets tile size and buffer if not supplied.
}
