% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_tiling.R
\name{perform_tiling}
\alias{perform_tiling}
\title{Perform Tiling on LAS files}
\usage{
perform_tiling(
  proj_dir,
  tile_size = 250,
  chunk_buffer = 0,
  index_tiles = TRUE,
  output_laz = TRUE,
  num_cores = 1L,
  acq = NULL
)
}
\arguments{
\item{proj_dir}{Project directory containing the raw las files.}

\item{tile_size}{Size of each tile in meters.}

\item{chunk_buffer}{Buffer area around each tile.}

\item{index_tiles}{Logical, should tiles be indexed?}

\item{output_laz}{Logical, should the output be in LAZ format?}

\item{num_cores}{Number of cores for parallel processing.}

\item{acq}{Acquisition name.}
}
\value{
A catalog of tiled LAS files.
}
\description{
Perform Tiling on LAS files
}
