% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_load_file.R, R/select_file_path.R
\name{select_file_path}
\alias{select_file_path}
\title{Select file path from directory}
\usage{
select_file_path(file_dir, pattern, ext = ".tif$")

select_file_path(file_dir, pattern, ext = ".tif$")
}
\arguments{
\item{file_dir}{A string specifying the directory to check.}

\item{pattern}{A string specifying the pattern to match file names against.}

\item{ext}{A string specifying the file extension to filter on. Default is '.tif$'.}
}
\value{
A string containing the full path of the selected file.

A string containing the full path of the selected file.
}
\description{
This function checks a specified directory for files matching a specified pattern,
and either loads a file if there's only one match, or gives the user an option
to choose which file to load if there are multiple matches.
If no files match the specified pattern, an error is returned.

This function checks a specified directory for files matching a specified pattern,
and either loads a file if there's only one match, or gives the user an option
to choose which file to load if there are multiple matches.
If no files match the specified pattern, an error is returned.
}
\examples{
select_file_path("path/to/directory", "pattern")
select_file_path("path/to/directory", "pattern")
}
