% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_als_dirs.R
\name{setup_als_dirs}
\alias{setup_als_dirs}
\title{Setup standard directories for an LAS project}
\usage{
setup_als_dirs(proj_dir)
}
\arguments{
\item{proj_dir}{A character string representing the path to the root directory of the ALS project.}
}
\value{
A list of created directories, invisibly.
}
\description{
This function creates a set of standardized directories for managing and organizing
Airborne LiDAR (ALS) or other LAS data and associated spatial data, including input, output,
documentation, and script folders.
}
\examples{
# Set up directories for an example ALS project
proj_path <- "path/to/your/project"
setup_als_dirs(proj_path)

}
