% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silvtools-package.R
\docType{package}
\name{silvtools-package}
\alias{silvtools}
\alias{silvtools-package}
\title{silvtools: Handy Functions for Tree Level Analyses}
\description{
This package is my first attempt and focusing on reproducible code for processing involving high density airborne lidar and other remotely datasets used in my PhD project focused on fine-scale remote sensing to inform mid-rotation silvicultural management and research.
}
\author{
\strong{Maintainer}: Liam Irwin \email{liamakirwin@gmail.com}

}
\keyword{internal}
