% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_simulator.R
\name{solar_simulator}
\alias{solar_simulator}
\title{Title}
\usage{
solar_simulator(
  dsm_file,
  weekday = "Mon",
  time_df,
  proj_dir,
  site_name,
  aoi_file = NULL,
  sun_pos,
  interval,
  start_time,
  end_time,
  filt = FALSE
)
}
\arguments{
\item{filt}{}
}
\description{
Title
}
