% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_matching.R
\name{tree_matching}
\alias{tree_matching}
\title{Match ground truth with detected tree tops
Calculates distance between ground truth and nearest 2 detected tree tops, returns a matched dataset containing the
tree ID of the ground truth and detected matches}
\usage{
tree_matching(reference, detected)
}
\arguments{
\item{reference}{sf object - points of recorded tree locations gathered in field}

\item{detected}{sf object - tree top locations typically from lidR::locate_trees}
}
\value{
returns a tree matching data table
}
\description{
Match ground truth with detected tree tops
Calculates distance between ground truth and nearest 2 detected tree tops, returns a matched dataset containing the
tree ID of the ground truth and detected matches
}
\examples{
\dontrun{
detected <- lidR::locate_trees(las, algorithm = lmf(ws = 2, hmin = 10))

reference <- sf::st_read('F:/Quesnel_2022/Quesnel_2022_PosTex/
georeferenced_adjusted/CT1P1_postex_utm10n.shp')

matched_trees <- tree_matching(reference, detected)
}
}
