% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_matching_scores.R
\name{tree_matching_scores}
\alias{tree_matching_scores}
\title{Calculate Tree Matching Scores}
\usage{
tree_matching_scores(match_table)
}
\arguments{
\item{match_table}{output from silvtools::tree_matching}
}
\value{
returns dataframe with metrics including number of FP/TP/FN as well as
precision, recall, Fscore, and statistics on distances between matches
}
\description{
Calculate Tree Matching Scores
}
\examples{
\dontrun{
reference <- st_read('ground_reference_stems.shp')
detected <- locate_trees(chm, lmf(ws = 2))
match_table <- tree_matching(reference, detected)
scores <- tree_matching_scores(match_table)
}
}
