% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape_metrics.R
\name{alphashape_metrics}
\alias{alphashape_metrics}
\title{Generate Crown Structural Metrics}
\usage{
alphashape_metrics(las)
}
\arguments{
\item{las}{single tree las object}
}
\value{
returns a dataframe with 3D convex hull metrics
}
\description{
Generate Crown Structural Metrics
}
\examples{
\dontrun{
las <- lidR::filter_poi(tree_las, treeID == 17)
ashape <- alphashape_metrics(las)
}
}
