% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_chm_crowns.R
\name{approximate_chm_crowns}
\alias{approximate_chm_crowns}
\title{Segment tree crowns from a CHM}
\usage{
approximate_chm_crowns(
  proj_dir,
  chm_dir = NULL,
  vector_output = NULL,
  raster_output = NULL,
  crown_methods = c("fixed", "auto", "variable"),
  hmin = 1,
  crown_height_threshold = 0.25,
  vis = FALSE,
  chm_ext = "smooth",
  window_size = NULL,
  lmf_ext = NULL
)
}
\arguments{
\item{proj_dir}{A string specifying the project directory.}

\item{chm_dir}{A string specifying the raster output directory. Default is NULL.}

\item{vector_output}{A string specifying the vector output directory. Default is NULL.}

\item{raster_output}{A string specifying the raster output directory. Default is NULL.}

\item{crown_methods}{A character vector specifying the crown segmentation methods to use. Default is c('fixed', 'auto', 'variable').}

\item{hmin}{A numeric specifying the minimum height of a tree. Default is 1.}

\item{crown_height_threshold}{A numeric specifying the threshold for crown height. Default is 0.25.}

\item{vis}{A logical specifying whether to visualize the segmentation results. Default is FALSE.}

\item{chm_ext}{A string specifying the extension of the CHM file. Default is 'smooth'.}

\item{window_size}{A numeric specifying the window size for the segmentation method. Default is NULL.}

\item{lmf_ext}{A string specifying the custom pattern for tree top files. Default is 'lmf_ws_{window_size}'.}
}
\value{
A list containing the results of each segmentation method.
}
\description{
This function segments tree crowns using a canopy height model (CHM) based on the tree tops located earlier.
}
\examples{
\dontrun{
approximate_chm_crowns("path/to/project")
}
}
