% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_mask.R
\name{crown_mask}
\alias{crown_mask}
\title{Watershed segmentation of crowns with threshold}
\usage{
crown_mask(
  chunk,
  chm_res = 0.25,
  chm_cutoff_percent = 0.5,
  ws = 2,
  uniqueness = "incremental",
  vis = FALSE
)
}
\arguments{
\item{chunk}{lasCatalog chunk (catalog_apply) or las file loaded with readLAS}

\item{chm_res}{desired resolution of chm for segmentation}

\item{chm_cutoff_percent}{desired threshold of treetop height to mask crown}

\item{ws}{local maxima finding window size}

\item{uniqueness}{see ?lidR::locate_trees argument defines how tree IDs are determined}

\item{vis}{LOGICAL; choose whether or not to plot result (default = FALSE)}
}
\description{
Watershed segmentation of crowns with threshold
}
