% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chm.R
\name{generate_chm}
\alias{generate_chm}
\title{Generate Canopy Height Model (CHM)}
\usage{
generate_chm(
  proj_dir,
  res = 1,
  algorithm = c("p2r", "pitfree"),
  subcircle_p2r = 0,
  subcircle_pitfree = 0,
  num_cores = 1L,
  chunk_buf = NULL,
  acq = NULL,
  clamp = FALSE,
  min_clamp = 0,
  max_clamp = 50
)
}
\arguments{
\item{proj_dir}{Project directory containing the normalized LAS files.}

\item{res}{CHM resolution in meters. Default is 1.}

\item{algorithm}{Algorithm to use for CHM generation. Default is 'p2r'. Options are 'p2r' and 'pitfree'}

\item{subcircle_p2r}{Subcircle argument for p2r CHM generation, default is 0.}

\item{subcircle_pitfree}{Subcircle argument for pitfree CHM generation, default is 0.}

\item{num_cores}{Number of cores for parallel processing. Default is 1.}

\item{chunk_buf}{Chunk buffer size in meters. Default is NULL (auto-set to 5\% of tile_size).}

\item{acq}{Acquisition name. Default is NULL.}

\item{clamp}{Logical switch to clamp CHM values. Default is FALSE.}

\item{min_clamp}{Minimum value to clamp CHM at. Default is 0.}

\item{max_clamp}{Maximum value to clamp CHM at. Default is 50.}
}
\value{
Saves a CHM locally as well as smoothed and filled versions of the CHM.
}
\description{
This function generates a Canopy Height Model (CHM) from normalized LAS files.
}
