% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dtm.R
\name{generate_dtm}
\alias{generate_dtm}
\title{Generate Digital Terrain Model (DTM)}
\usage{
generate_dtm(
  proj_dir,
  res = 1,
  dtm_algorithm = "tin",
  num_cores = 1L,
  chunk_buf = NULL,
  k = 8,
  acq = basename(proj_dir)
)
}
\arguments{
\item{proj_dir}{Project directory containing the classified LAS files.}

\item{res}{DTM resolution in meters.}

\item{dtm_algorithm}{Custom DTM algorithm. Default is NULL (TIN). c('tin', 'knnidw', 'kriging')}

\item{num_cores}{Number of cores for parallel processing. Default is 1.}

\item{chunk_buf}{Chunk buffer size in meters. Default is NULL (auto-set to 5\% of tile_size).}

\item{k}{Number of neighbours of IDW/KNN. Default is 8}

\item{acq}{Acquisition name. Default is NULL.}
}
\value{
Saves DTMs locally to proj_dir/output/raster/dtm
}
\description{
This function generates a Digital Terrain Model (DTM) from classified LAS files.
}
