% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heygi_cindex.R
\name{heygi_cindex}
\alias{heygi_cindex}
\title{Calculate Heygi Style Competition Indicies}
\usage{
heygi_cindex(ttops, comp_input = "vol_convex", maxR = 6)
}
\arguments{
\item{ttops}{tree tops object; expects an X.detected, Y.detected column, typically from silvtools::tree_matching() function}

\item{comp_input}{string value indicating competition index input; traditionally DBH/Height, we test crown structural metrics}

\item{maxR}{radius of sphere of influence; treetops within this radius will be included in the calculation of cindex}
}
\value{
returns original ttops dataset (sf/dataframe) with new cindex column
}
\description{
Calculate Heygi Style Competition Indicies
}
\examples{
\dontrun{
c <- heygi_cindex(ttops, comp_input = 'vol_convex', maxR = 6)
}
}
