% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_height_normalization.R
\name{perform_height_normalization}
\alias{perform_height_normalization}
\title{Perform Height Normalization on Classified LAS Files}
\usage{
perform_height_normalization(
  proj_dir,
  chunk_buf = NULL,
  tile_size = NULL,
  acq = NULL,
  norm_algorithm = "tin",
  num_cores = 1L,
  index_tiles = TRUE,
  output_laz = TRUE
)
}
\arguments{
\item{proj_dir}{Project directory containing the classified las files.}

\item{chunk_buf}{Buffer area around each tile in meters.}

\item{tile_size}{The size of each tile for processing. Default is NULL, which will use the size of the middle tile in the catalog.}

\item{acq}{Acquisition name. Default is NULL.}

\item{norm_algorithm}{Normalization algorithm to use. Default is tin().}

\item{num_cores}{Number of cores for parallel processing. Default is 1.}

\item{index_tiles}{Logical, should tiles be indexed? Default is TRUE.}

\item{output_laz}{Logical, should the output be in LAZ format? Default is TRUE.}
}
\value{
A catalog of height-normalized LAS files.
}
\description{
Perform Height Normalization on Classified LAS Files
}
