% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_simulator.R
\name{solar_simulator}
\alias{solar_simulator}
\title{Rayshading solar simulator function}
\usage{
solar_simulator(
  dsm_file,
  proj_dir,
  site_name,
  aoi_file = NULL,
  sun_pos,
  num_cores = NULL
)
}
\arguments{
\item{dsm_file}{File path to DSM file of interest typically 'path/dsm.tif'}

\item{proj_dir}{Directory for block/acquisition creates an /output/irradiance folder within proj_dir}

\item{site_name}{Site name for file naming purposes ex 'CT1'}

\item{aoi_file}{Optional area of interest for internal cropping of DSM either file pointing to shp/gpkg etc or loaded sf object}

\item{sun_pos}{Sun position data frame from get_sun_pos, zenith, azimuth for each timepoint of interest}

\item{num_cores}{Number of cores to use for parallel processing. Default is 1.}
}
\value{
Writes a raster for each timepoint recording the irradiance for each given pixel
}
\description{
Rayshading solar simulator function
}
\examples{
\dontrun{

# Example of running solar simulator on a sample DSM

# Load sample Megaplot data
lasfile <- system.file("extdata", "Megaplot.laz", package = "lidR")
las <- readLAS(lasfile)

# Set up project directory and site name
proj_dir <- tempdir()
site_name <- 'megaplot_ex'

# Generate DSM and save to file
dsm <- rasterize_canopy(las, res = 5)
dsm_file <- file.path(proj_dir, paste0(site_name,'_dsm.tif'))
writeRaster(dsm, dsm_file)

# Set up solar position data
ext_sf <- ext(dsm) \%>\% vect(crs = crs(dsm)) \%>\% st_as_sf() \%>\% st_transform(4326)
coords <- st_coordinates(ext_sf)[1,]
lat <- as.numeric(coords[2])
lon <- as.numeric(coords[1])
start_date <- as.POSIXct("2022-09-15 08:00:00", tz = "America/Los_Angeles")
end_date <- as.POSIXct("2022-09-15 10:00:00", tz = "America/Los_Angeles")
interval <- '10 min'

# Grab sun position data
sun_pos <- get_solar_pos(start_date, end_date, interval, lat, lon)
# Perform solar simulation across timepoints
solar_simulator(dsm_file, proj_dir, site_name, aoi_file, sun_pos)
}
}
